+++ src/index.ts
@@ -1,3 +1,4 @@
 export * from './models';
 export * from './design-values';
 export * from './themes';
+export * from './storage';